+++ tests/test_core_to_contrib.py
@@ -1309,6 +1309,10 @@ OPERATOR = [
         'airflow.providers.amazon.aws.operators.hive_to_dynamodb.HiveToDynamoDBTransferOperator',
         'airflow.contrib.operators.hive_to_dynamodb.HiveToDynamoDBTransferOperator',
     ),
+    (
+        'airflow.providers.amazon.aws.operators.imap_attachment_to_s3.ImapAttachmentToS3Operator',
+        'airflow.contrib.operators.imap_attachment_to_s3_operator.ImapAttachmentToS3Operator',
+    ),
 ]
 
 SENSOR = [
